hrms_project/requirements.txt
django
django-cors-headers
djangorestframework
gunicorn
whitenoise
mysqlclient
